requirements.txt
requests
kafka-python
python-dotenv
boto3
pyspark==3.5.4
